Run on akran (Pg 11, Ubuntu 18, same hw as claudrin)
Plain inserts are noticably slower than they were on claudrin, copy and
select are faster. Not sure if that is due to Pg 11, the fact that this
is a cluster, the different OS or something else.

import_execute_batch is about 50% faster than plain insert, but still
much slower than copy. The page size doesn't seem to make much
difference. 10000 (used in this run) is faster than the default of 100,
but 1000 was actually slower.